+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## 01-05-2025, v 0.9.1
+- Updated to GPT Image 1
+- Removed Image URL downloader
+- Updated prompt system to allow for more creative story generation
+- Updated Response formats to be more consistent with SFX and other panel text components
+
 ## 22-09-2023, v 0.9
 - First Commit